feat(TokenWizard): create a new rule to compute the weight f a relation
Because we are using louvain partition method, we need to make sure not to create to high weights on an edges. Because if we do so, the edge is going to be to strong and won't cut during the partition step. SO we decided to lower off by 20% the weight of a relation when it is the same tokens are compared.
In the next commit, you'll see that the results have improved